+++ src/detail/http_client.cpp
@@ -41,10 +41,9 @@ nlohmann::json HttpClient::GetJson(const std::string& path,
 }
 
 nlohmann::json HttpClient::PostJson(const std::string& path,
-                                    const httplib::Params& params) {
-  // need to fully specify, otherwise sent as application/x-www-form-urlencoded
-  const std::string full_path = httplib::append_query_params(path, params);
-  httplib::Result res = client_.Post(full_path);
+                                    const httplib::Params& form_params) {
+  // params will be encoded as form data
+  httplib::Result res = client_.Post(path, {}, form_params);
   return HttpClient::CheckAndParseResponse(path, std::move(res));
 }